Updated function doc-string
The doc-string information for `compute_pearson_multi_network` contained an error, in which all the columns were computed/populated all the way to `n` - which was not the case. The documentation was corrected to be consistent with what was actually being done - which was constructing a N x N square symmetric matrix from matrices with the same N number of rows but different number of columns.